bin: Add a silent option to onenote-parse
Make printing the parsed structures configurable, so that one could
infer by the exit status (and possibly the error traces) if something
has gone wrong, to make development a bit easier.
This does not change the default behavior of the binary, so it will
still print without an option supplied to it.