fsshttpb: packaging: Add a CellSchemaId enum and use it for file types
Refactor some open-coded uses of GUIDs when determining what file we're
up to, and define them in the new determine_format() function in
OneStorePackaging.
Note that we compared reference packaging .one and .onetoc2 GUIDs
against the value returned by store.schema_guid(), which referred to the
storage manifest schema GUID. Now, we do it against the GUID in the
packaging format header cell. These seem to be equal, but I'm not 100%
sure.